fix(fcbc): correct opcode width and table type in the new tests
The first Full Gate run caught two compile errors the local static
review missed: `ExpressionNode.opcode` is `u16` (the `query` helper took
`u32`), and the scaling-case table type tripped
`clippy::type_complexity` (now behind a `ScalingCase` alias).

+++ crates/fcs-fcbc/src/evaluator.rs
@@ -1324,7 +1324,7 @@ mod tests {
             node
         };
         let query = |chart: &mut DecodedChart,
-                     opcode: u32,
+                     opcode: u16,
                      result_type: ValueType,
                      left: u32,
                      right: u32| {
